+++ src/pipeline/build.ts
@@ -20,7 +20,7 @@ import { compile } from "./compile";
 import { precompile } from "./precompile";
 import { getCompilerVersion } from "./version";
 import { FactoryAst, getAstFactory, idText } from "../ast/ast";
-import { TactErrorCollection } from "../error/errors";
+import { TactError, TactErrorCollection } from "../error/errors";
 import { getParser, Parser } from "../grammar";
 import { defaultParser } from "../grammar/grammar";
 
@@ -83,7 +83,13 @@ export async function build(args: {
                 ? "Syntax and type checking failed"
                 : "Tact compilation failed",
         );
-        logger.error(e as Error);
+
+        // show an error with a backtrace only in verbose mode
+        if (e instanceof TactError && config.verbose && config.verbose < 2) {
+            logger.error(e.message);
+        } else {
+            logger.error(e as Error);
+        }
         return { ok: false, error: [e as Error] };
     }
 
@@ -146,7 +152,16 @@ export async function build(args: {
             codeEntrypoint = res.output.entrypoint;
         } catch (e) {
             logger.error("Tact compilation failed");
-            logger.error(e as Error);
+            // show an error with a backtrace only in verbose mode
+            if (
+                e instanceof TactError &&
+                config.verbose &&
+                config.verbose < 2
+            ) {
+                logger.error(e.message);
+            } else {
+                logger.error(e as Error);
+            }
             ok = false;
             errorMessages.push(e as Error);
             continue;